fix(search): 인기태그로 검색 진입 시 검색창에 키워드 보이도록 변경
