+++ src/index.tsx
@@ -4,10 +4,9 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import App from './App';
-// import reportWebVitals from './reportWebVitals';
 import HomePage from './pages/HomePage';
 import FollowingPage from './pages/FollowingPage';
-import SearchPage from './pages/SearchPage';
+import SearchPage from './pages/search/SearchPage';
 import MyPage from './pages/mypage/MyPage';
 import WishPage from './pages/WishPage';
 import AlarmPage from './pages/AlarmPage';
